# static/openapi.yaml
openapi: 3.1.0
info:
  title: API for math team competition
  version: 0.0.1
...
paths:
...
  /api/contests:
    get:
      tags:
        - contests
      summary: Get a list of contests
      description: Retrieves the list of contests
      responses:
        "200":
          description: A list of contests was retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  contests:
                    type: array
                    items:
                      type: integer
                      description: The ID of a contest
        "401":
          description: The user was not authenticated
    post:
      tags:
        - contests
      summary: Create a new contest
      description: Creates a new contest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phiquadro_id:
                  type: number
                  minimum: 1
                  description: The ID of the contest on PhiQuadro
                phiquadro_sess:
                  type: number
                  minimum: 1
                  description: The session of the contest on PhiQuadro
                name:
                  type: string
                  description: The name of the new contest
                start_time:
                  type: string
                  description: The start time of the new contest
                  example: 2025-01-30T10:10
                duration:
                  type: integer
                  minimum: 1
                  description: The duration of the new contest, in minutes
                drift:
                  type: integer
                  minumum: 0
                  description: The drift of the new contest
                drift_time:
                  type: integer
                  minimum: 0
                  description: The duration of the drift, in minutes
                jolly_time:
                  type: integer
                  minimum: 0
                  description: The time to choose a jolly, in minutes
                question_bonus:
                  type: array
                  items:
                    type: integer
                  minItems: 10
                  maxItems: 10
                  description: The bonus awarded for the first solution to a question
                contest_bonus:
                  type: array
                  items:
                    type: integer
                  minItems: 10
                  maxItems: 10
                  description: The bonus awarded for the first solution to all questions
      responses:
        "201":
          description: The contest was created
          content:
            application/json:
              schema:
                type: object
                properties:
                  contest_id:
                    type: number
                    minimum: 1
                    description: The ID of the contest
        "400":
          description: The request was malformed
        "401":
          description: The user was not authenticated
        "422":
          description: The request had semantic errors
  /api/contests/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: The ID of the contest
    get:
      tags:
        - contests
      summary: Retrieves a contest
      responses:
        "200":
          description: A list of contest is being provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  phiquadro_id:
                    type: number
                    description: The ID of the contest on phiquadro
                  phiquadro_sess:
                    type: number
                    description: The session of the contest on phiquadro
                  name:
                    type: string
                    description: The name of the contest
                  start_time:
                    type: string
                    description: The start time of the contest
                    example: 2025-01-30T10:10
                  duration:
                    type: number
                    description: The duration of the contest in seconds
                  drift:
                    type: number
                    description: The drift of the contest
                  drift_duration:
                    type: number
                    description: The duration of the drift of the contest in seconds
                  jolly_time:
                    type: integer
                    minimum: 0
                    description: The time to choose a jolly, in minutes
                  question_bonus:
                    type: array
                    items:
                      type: integer
                    minItems: 10
                    maxItems: 10
                    description: The bonus awarded for the first solution to a question
                  contest_bonus:
                    type: array
                    items:
                      type: integer
                    minItems: 10
                    maxItems: 10
                    description: The bonus awarded for the first solution to all questions
        "401":
          description: The user was not authenticated
        "404":
          description: The requested contest was not found or is not accessible to the user
    patch:
      tags:
        - contests
      summary: Applies changes to the contest, only modifying present fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                start_time:
                  type: string
                  description: The start time of the contest
                  example: 2025-01-30T10:10
                duration:
                  type: integer
                  minimum: 1
                  description: The duration of the new contest, in minutes
                drift:
                  type: integer
                  minumum: 0
                  description: The drift of the new contest
                drift_time:
                  type: integer
                  minimum: 0
                  description: The duration of the drift, in minutes
                jolly_time:
                  type: integer
                  minimum: 0
                  description: The time to choose a jolly, in minutes
                question_bonus:
                  type: array
                  items:
                    type: integer
                  minItems: 10
                  maxItems: 10
                  description: The bonus awarded for the first solution to a question
                contest_bonus:
                  type: array
                  items:
                    type: integer
                  minItems: 10
                  maxItems: 10
                  description: The bonus awarded for the first solution to all questions
      responses:
        "204":
          description: The update was successful
        "400":
          description: The request was malformed
        "401":
          description: The user was not authenticated
        "403":
          description: The user requested to perform an action they were not allowed to perform
        "404":
          description: The requested contest was not found or is not accessible to the user
        "422":
          description: The request has semantic errors
    delete:
      tags:
        - contests
      summary: Deletes the contest
      responses:
        "204":
          description: The deletion was successful
        "401":
          description: The user was not authenticated
        "404":
          description: The requested contest was not found or is not accessible to the user
...